fix lint, fix tessts
