+++ e2e/commands/eject.e2e.1.ts
@@ -16,6 +16,7 @@ describe('bit eject command', function () {
   let helper: Helper;
   before(() => {
     helper = new Helper();
+    helper.command.setFeatures('legacy-workspace-config');
   });
   describe('local component', () => {
     before(() => {
