+++ e2e/functionalities/dynamic-namespaces.e2e.3.ts
@@ -10,6 +10,7 @@ describe('dynamic namespaces', function () {
   let helper: Helper;
   before(() => {
     helper = new Helper();
+    helper.command.setFeatures('legacy-workspace-config');
   });
   after(() => {
     helper.scopeHelper.destroy();
