+++ e2e/functionalities/repository-hooks.e2e.ts
@@ -12,6 +12,7 @@ describe('repository-hooks', function () {
   let helper: Helper;
   before(() => {
     helper = new Helper();
+    helper.command.setFeatures('legacy-workspace-config');
   });
   after(() => {
     helper.scopeHelper.destroy();
