+++ e2e/harmony/add-harmony.e2e.ts
@@ -1,5 +1,6 @@
-import chai from 'chai';
+import chai, { expect } from 'chai';
 import path from 'path';
+import { AddingIndividualFiles } from '../../src/consumer/component-ops/add-components/exceptions/adding-individual-files';
 import { ParentDirTracked } from '../../src/consumer/component-ops/add-components/exceptions/parent-dir-tracked';
 import Helper from '../../src/e2e-helper/e2e-helper';
 
@@ -13,6 +14,22 @@ describe('add command on Harmony', function () {
   });
   after(() => {
     helper.scopeHelper.destroy();
+  });
+  describe('adding files when workspace is new', () => {
+    before(() => {
+      helper.scopeHelper.setNewLocalAndRemoteScopesHarmony();
+      helper.fixtures.createComponentBarFoo();
+    });
+    it('should throw an error AddingIndividualFiles', () => {
+      const addFunc = () => helper.command.addComponent('bar/foo.js');
+      const error = new AddingIndividualFiles(path.normalize('bar/foo.js'));
+      helper.general.expectToThrow(addFunc, error);
+    });
+    it('when excluding a file, should throw an error', () => {
+      helper.fs.outputFile('bar/foo1.js');
+      const cmd = () => helper.command.addComponent('bar', { e: 'bar/foo1.js' });
+      expect(cmd).to.throw('unable to exclude files when tracking a directory');
+    });
   });
   describe('add a directory inside an existing component', () => {
     before(() => {
